feat: refactor preshift for better organization
